+++ example/lib/gallery/sections/icons.dart
@@ -470,6 +470,16 @@ class IconsSection extends StatelessWidget {
                         is16x16: true,
                       ),
                       _IconEntry(
+                        data: NesIcons16.topArrowIndicator,
+                        label: 'topArrowIndicator',
+                        is16x16: true,
+                      ),
+                      _IconEntry(
+                        data: NesIcons16.bottomArrowIndicator,
+                        label: 'bottomArrowIndicator',
+                        is16x16: true,
+                      ),
+                      _IconEntry(
                         data: NesIcons16.bell,
                         label: 'bell',
                         is16x16: true,